requirements.txt
astroid==2.3.3
autopep8==1.5.1
boto3==1.12.39
botocore==1.15.39
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
docker==4.2.0
docutils==0.15.2
dynamodb-json==1.3
Flask==1.1.2
Flask-Cors==3.0.8
haversine==2.2.0
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
onesignal-sdk==1.1.0
pycodestyle==2.5.0
pylint==2.4.4
python-dateutil==2.8.1
requests==2.23.0
s3transfer==0.3.3
simplejson==3.17.0
six==1.14.0
typed-ast==1.4.1
urllib3==1.25.8
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.11.2
